+++ config/routes.rb
@@ -23,6 +23,7 @@ Rails.application.routes.draw do
   resources :generations, path: 'results', only: %i[index show create destroy] do
     member do
       post :retry
+      post :cancel
       patch :share, action: :update_share
     end
   end